Moving listener to other namespace level
