+++ tests/src/EventManager/Listener/CollectionTest.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace PHPFluent\EventManager;
+namespace PHPFluent\EventManager\Listener;
 
 use ReflectionFunction;
 
 /**
  * @covers PHPFluent\EventManager\ListenerCollection
  */
-class ListenerCollectionTest extends \PHPUnit_Framework_TestCase
+class CollectionTest extends \PHPUnit_Framework_TestCase
 {
     public function testShouldAddListenerToCollection()
     {
         $listener   = $this->getMock('PHPFluent\EventManager\Listener');
-        $collection = new ListenerCollection();
+        $collection = new Collection();
 
         $collection->add($listener);
 
@@ -21,7 +21,7 @@ class ListenerCollectionTest extends \PHPUnit_Framework_TestCase
 
     public function testShouldBeIterable()
     {
-        $collection = new ListenerCollection();
+        $collection = new Collection();
 
         $collection->add($this->getMock('PHPFluent\EventManager\Listener'));
